test(n-vm): proper integration tests
No need for scratch main to run this type of test.
