--- scratch/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "scratch"
-version = "0.0.0"
-edition = "2024"
-license = "Apache-2.0"
-publish = false
-
-[dependencies]
-
-# internal
-n-vm = { path = "../n-vm" }
-
-tokio = { workspace = true, default-features = false, features = ["rt", "process", "macros", "fs", "net", "time", "io-util", "net", "io-std"] }
-tracing-subscriber = { workspace = true, default-features = false, features = ["fmt"] }
-tracing = { workspace = true, default-features = false, features = [] }
